rename oss-fuzz project and adopt latest formatting updates. Just to test build
