+++ projects/pysecsan/pysecsan/tests/poe/libvcs-cve-2022-21187/fuzz_libvcs.py
@@ -12,31 +12,34 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Targets: https://github.com/python-ldap/python-ldap/security/advisories/GHSA-r8wq-qrxc-hmcm"""
+"""Targets https://github.com/advisories/GHSA-mv2w-4jqc-6fg4"""
 
 import os
 import sys
 import atheris
-
 import pysecsan
+
 pysecsan.add_hooks()
 
-import ldap.schema
+from libvcs.shortcuts import create_repo
 
 
 def TestOneInput(data):
-    fdp = atheris.FuzzedDataProvider(data)
-    try:
-        ldap.schema.split_tokens(fdp.ConsumeUnicodeNoSurrogates(1024))
-    except ValueError:
-        pass
+  fdp = atheris.FuzzedDataProvider(data)
+  mercurial_repo = create_repo(url=fdp.ConsumeUnicodeNoSurrogates(128),
+                               vcs="hg",
+                               repo_dir="./")
+  try:
+    mercurial_repo.update_repo()
+  except (ValueError, FileNotFoundError) as e:
+    pass
 
 
 def main():
-    atheris.instrument_all()
-    atheris.Setup(sys.argv, TestOneInput, enable_python_coverage=True)
-    atheris.Fuzz()
+  atheris.instrument_all()
+  atheris.Setup(sys.argv, TestOneInput, enable_python_coverage=True)
+  atheris.Fuzz()
 
 
 if __name__ == "__main__":
-    main()
+  main()
