+++ projects/pysecsan/pysecsan/tests/poe/python-ldap-GHSL-2021-117/fuzz_ldap.py
@@ -12,45 +12,32 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+"""Targets: https://github.com/python-ldap/python-ldap/security/advisories/GHSA-r8wq-qrxc-hmcm"""
 
 import os
 import sys
 import atheris
 
 import pysecsan
-import yaml
 
-def serialize_with_tainted_data(param, magicval):
-    if magicval == 1337:
-        try:
-            yaml.load(param, yaml.Loader)
-        except yaml.YAMLError:
-            pass
-    elif magicval == 1338:
-        try:
-            yaml.load("FROMFUZZ", yaml.Loader)
-        except yaml.YAMLError:
-            pass
-    else:
-        return 2
+pysecsan.add_hooks()
+
+import ldap.schema
 
 
 def TestOneInput(data):
-    fdp = atheris.FuzzedDataProvider(data)
-    serialize_with_tainted_data(
-        fdp.ConsumeUnicodeNoSurrogates(32),
-        fdp.ConsumeIntInRange(500, 1500)
-    )
+  fdp = atheris.FuzzedDataProvider(data)
+  try:
+    ldap.schema.split_tokens(fdp.ConsumeUnicodeNoSurrogates(1024))
+  except ValueError:
+    pass
 
 
 def main():
-    pysecsan.add_hooks()
-
-    atheris.instrument_all()
-    atheris.Setup(sys.argv, TestOneInput, enable_python_coverage=True)
-    atheris.Fuzz()
+  atheris.instrument_all()
+  atheris.Setup(sys.argv, TestOneInput, enable_python_coverage=True)
+  atheris.Fuzz()
 
 
 if __name__ == "__main__":
-    main()
-
+  main()
